requirements.txt
boto3
fastapi
numpy
opencv-python-headless
pillow
python-dotenv
pydantic-settings
requests
torch
torchvision
uvicorn[standard]
